リファクタリング。ロジックのClass化と階層構造化。
