+++ src/main/resources/static/index.html
@@ -5,7 +5,7 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="screanId" content="main">
+    <meta name="screenId" content="main">
     <title>Twitter Buttler - メインメニュー</title>
 
     <!-- Bootstrap -->
